Fixed bugreport:5699 all new options should now be passed properly to clones and mobs.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@16132 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/mob.h
@@ -104,7 +104,7 @@ struct mob_db {
 	struct { int nameid,p; } mvpitem[MAX_MVP_DROP];
 	struct status_data status;
 	struct view_data vd;
-	short option;
+	unsigned int option;
 	int summonper[MAX_RANDOMMONSTER];
 	int maxskill;
 	struct mob_skill skill[MAX_MOBSKILL];
